Move npm dependencies checksum calculation to build-info-go
