+++ artifactory/commands/npm/common.go
@@ -32,10 +32,9 @@ type CommonArgs struct {
 	// Npm registry as exposed by Artifactory.
 	registry string
 	// Npm token generated by Artifactory using the user's provided credentials.
-	npmAuth         string
-	typeRestriction biutils.TypeRestriction
-	authArtDetails  auth.ServiceDetails
-	npmVersion      *version.Version
+	npmAuth        string
+	authArtDetails auth.ServiceDetails
+	npmVersion     *version.Version
 	NpmCommand
 }
 
@@ -114,29 +113,6 @@ func (com *CommonArgs) createTempNpmrc() error {
 	return errorutils.CheckError(ioutil.WriteFile(filepath.Join(com.workingDirectory, npmrcFileName), configData, 0600))
 }
 
-func (com *CommonArgs) setTypeRestriction(key string, value string) {
-	// From npm 7, type restriction is determined by 'omit' and 'include' (both appear in 'npm config ls').
-	// Other options (like 'dev', 'production' and 'only') are deprecated, but if they're used anyway - 'omit' and 'include' are automatically calculated.
-	// So 'omit' is always preferred, if it exists.
-	if key == "omit" {
-		if strings.Contains(value, "dev") {
-			com.typeRestriction = biutils.ProdOnly
-		} else {
-			com.typeRestriction = biutils.All
-		}
-	} else if com.typeRestriction == biutils.DefaultRestriction { // Until npm 6, configurations in 'npm config ls' are sorted by priority in descending order, so typeRestriction should be set only if it was not set before
-		if key == "only" {
-			if strings.Contains(value, "prod") {
-				com.typeRestriction = biutils.ProdOnly
-			} else if strings.Contains(value, "dev") {
-				com.typeRestriction = biutils.DevOnly
-			}
-		} else if key == "production" && strings.Contains(value, "true") {
-			com.typeRestriction = biutils.ProdOnly
-		}
-	}
-}
-
 // This func transforms "npm config list" result to key=val list of values that can be set to .npmrc file.
 // it filters out any nil value key, changes registry and scope registries to Artifactory url and adds Artifactory authentication to the list
 func (com *CommonArgs) prepareConfigData(data []byte) ([]byte, error) {
@@ -161,7 +137,6 @@ func (com *CommonArgs) prepareConfigData(data []byte) ([]byte, error) {
 				} else {
 					filteredConf = append(filteredConf, currOption, "\n")
 				}
-				com.setTypeRestriction(key, value)
 			} else if strings.HasPrefix(splitOption[0], "@") {
 				// Override scoped registries (@scope = xyz)
 				filteredConf = append(filteredConf, splitOption[0], " = ", com.registry, "\n")
